create convert to json method
